call global ctors
