+++ Demos/DynamicLoader/DynExec/main.cpp
@@ -29,20 +29,8 @@ extern int g_lib_var;
 
 const char* g_string = "Hello, World!\n";
 
-void local_dbgputstr(const char* str, int len)
-{
-    // return str[0] + len;
-    constexpr unsigned int function = Kernel::SC_dbgputch;
-    for (int i = 0; i < len; ++i) {
-        unsigned int result;
-        asm volatile("int $0x82"
-                     : "=a"(result)
-                     : "a"(function), "d"((u32)str[i])
-                     : "memory");
-    }
-}
-
 int libfunc();
+void local_dbgputstr(const char* str, int len);
 
 int main(int, char**)
 {
