+++ Libraries/LibELF/DynamicLoader.h
@@ -107,7 +107,7 @@ private:
     // Stage 2
     void do_relocations();
     void setup_plt_trampoline();
-    void relocate_got_plt();
+    // void relocate_got_plt();
     void call_object_init_functions();
 
     Optional<u32> lookup_symbol(const ELF::DynamicObject::Symbol&) const;
